backend/backend/requirements.txt
fastapi>=0.115.6
uvicorn>=0.34.0
python-dotenv>=1.0.1
requests>=2.32.3
openai>=1.59.3
pydantic>=2.10.4
starlette>=0.41.3
python-multipart>=0.0.6
gunicorn>=21.2.0
aiohttp>=3.9.1
